Add files via upload
Python code to work with a 2-D CVM consisting of 256 (16x16) nodes; A = on = 1 = x1; B = off = 0 = x2
- Select a pre-stored pattern, OR
- Randomly generate a pattern. 
The randomly-generated pattern is adjusted to bring x1 close to x2, if they start too far apart. 
Once these patterns are obtained, the program:
1) Finds the configuration variables - nearest neighbor, etc., and
2) Computes the thermodynamic quantities (entropy, enthalpy, and free energy). 
Note: This version works with the equiprobable case only; x1 = x2 = 0.5 (approximately, in the case of the randomly-generated patterns). Thus, the enthalpy per unit activation is 0. The enthalpy for pairwise interactions, indicated by the h-value, does not have to be 0.